Fix: Userprofile now displays a correct profile depending if you're user or photographer, serviceCard also displayed correctly
