--- src/components/PhotographerProfile/ProfileCard/index.jsx
@@ -1,1 +1,0 @@
-export { default } from "./ProfileCard";
